Making getDatesForDays a helper function
